+++ tests/unit/AbstractConfigTest.php
@@ -49,6 +49,9 @@ abstract class AbstractConfigTest extends TestCase
                 'ints' => [1, 'a', 2, 1.5],
                 'floats' => [1.5, 'a', 2.5, 3],
             ],
+            'maps' => [
+                'mixed' => ['a' => 1, 0 => 'x', 'b' => 2, 7 => 'y'],
+            ],
         ];
         yield [$config, 'a.a.a', 'string', 'Foo'];
         yield [$config, 'a.a.b', 'int', 23];
@@ -65,6 +68,8 @@ abstract class AbstractConfigTest extends TestCase
         yield [$config, 'lists.strings', 'stringList', ['foo', 'bar']];
         yield [$config, 'lists.ints', 'intList', [1, 2]];
         yield [$config, 'lists.floats', 'floatList', [1.5, 2.5]];
+        // stringKeyedArray drops integer-keyed elements, preserving the string keys
+        yield [$config, 'maps.mixed', 'stringKeyedArray', ['a' => 1, 'b' => 2]];
     }
 
     /**
@@ -93,6 +98,7 @@ abstract class AbstractConfigTest extends TestCase
         yield ['stringList'];
         yield ['intList'];
         yield ['floatList'];
+        yield ['stringKeyedArray'];
     }
 
     /**
@@ -121,6 +127,7 @@ abstract class AbstractConfigTest extends TestCase
         yield ['foo', 'stringList'];
         yield ['foo', 'intList'];
         yield ['foo', 'floatList'];
+        yield ['foo', 'stringKeyedArray'];
     }
 
     /**
@@ -147,6 +154,7 @@ abstract class AbstractConfigTest extends TestCase
         yield [['foo', 'bar'], 'stringList'];
         yield [[1, 2], 'intList'];
         yield [[1.5, 2.5], 'floatList'];
+        yield [['a' => 1, 'b' => 2], 'stringKeyedArray'];
     }
 
     public function testNullableGet(): void